Add comments, and re-org.

+++ constexpr.cpp
@@ -5,18 +5,23 @@ constexpr int triple(const int &x) {
   return 3 * x;
 }
 
-template<typename T, int N>
+template<typename T, size_t N>
 struct array {
   T value [N];
 };
 
-int triple_func(const int &x) {
+size_t triple_func(const size_t &x) {
   return 3 * x;
 }
 
+/**
+ * constexpr can be calculated at compile time, and thus
+ * the array declaration wouldn't have compile error, while
+ * if using non-constexpr function, it would have compile error
+ */
 TEST(CONST_EXPR_TEST, constexpr_func) {
   // this implies triple(c) is calculated at compile time.
-  const int c = 5;
+  const size_t c = 5;
   array<int, triple(c)> dummy;
   static_assert(15 == sizeof(dummy.value)/sizeof(int), "");
 
@@ -24,6 +29,11 @@ TEST(CONST_EXPR_TEST, constexpr_func) {
   // array<int, triple_func(c)> dummy2;
 }
 
+/**
+ * Another test of constexpr for variable. The mainpoint here
+ * is that static_assert would result in a compile error if
+ * it was not declared as constexpr.
+ */
 TEST(CONST_EXPR_TEST, constexpr_var) {
   constexpr double PI = 3.141592653589793;
   static_assert(3.141592653589793 == PI, "");
